redis refactoring to use pub/sub and new namespaces
Databases are matched to storage strategies: runtime, dynamic, storage
the namespace for different keys, for instance modules, is dash separated
(also matching zsh name convention) plus an index list holding them.

The first listener appears for dns queries to gource, so that part of the
code can be concurrently launched and can be removed from the dnscap plugin.