import modprobe from kmod-22
Kmod is so wrong: passes arguments to its monolithic binary only
via symlinks to it, an incredibly convoluted way to add code that
we really don't need.

Sup needs to enforce suid execution of real files, not symlinks.

So we import modprobe with minimalist patches now and also the
syslog library.